requirements.txt
Flask
pymongo
python-dotenv
Flask-Login
google-generativeai
Markdown
beautifulsoup4